src/manager/pyproject.toml
[project]
name = "manager-django"
version = "0.1.0"
description = "Dockerized Django project using uv and Docker Compose."
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "psycopg[binary]",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "dj-database-url",
    "djangorestframework",
    "python-dotenv",
    "black",
    "isort"
]

[tool.uv]
# uv-specific configuration (if needed)
